requirements.txt
# Core
python-dotenv>=1.0.0  # Gestion des variables d'environnement
pyyaml>=6.0.1  # Parsing YAML
pydantic>=2.0.0  # Validation de données avec support Python 3.12+
pydantic-settings>=2.0.0  # Gestion des paramètres

# Logging et Monitoring
loguru>=0.7.0  # Logging structuré et coloré
structlog>=22.1.0
coloredlogs>=15.0.1  # Logs colorés
prometheus-client>=0.17.0  # Métriques
sentry-sdk>=1.30.0  # Monitoring des erreurs

# Données et Analyse
pandas>=2.0.0  # Manipulation de données
numpy>=1.26.0  # Calcul numérique
polars>=0.19.0  # DataFrame haute performance
pyarrow>=12.0.0  # Format de données colonnaire
pandas-ta>=0.3.14  # Indicateurs techniques supplémentaires
TA-Lib>=0.4.24  # Bibliothèque d'analyse technique

# Machine Learning et IA
scikit-learn>=1.4.0  # Algorithmes de ML
xgboost>=2.0.0  # Boosting d'arbres
tensorflow>=2.16.1  # Deep Learning
torch>=2.3.0  # PyTorch
optuna>=3.3.0  # Optimisation des hyperparamètres

# API et Communication
ccxt>=4.0.0  # API crypto-exchanges unifiée
krakenex>=2.2.2  # API Kraken native
fastapi>=0.110.0  # API REST moderne
uvicorn>=0.21.1  # Serveur ASGI
websockets>=12.0  # WebSockets pour le temps réel
httpx>=0.25.0  # Client HTTP asynchrone

# Base de Données
sqlalchemy>=2.0.12  # ORM SQL
asyncpg>=0.29.0  # Pilote PostgreSQL asynchrone
alembic>=1.12.0  # Migrations de base de données
redis>=5.0.1  # Cache et Pub/Sub
timescaledb>=2.11.0  # Extension PostgreSQL pour les séries temporelles

# Utilitaires
python-dateutil>=2.8.2  # Manipulation des dates
pytz>=2023.3  # Fuseaux horaires
tqdm>=4.66.0  # Barres de progression
pyjwt>=2.8.0  # JSON Web Tokens
python-multipart>=0.0.6  # Gestion des formulaires multipart

# Visualisation
matplotlib>=3.8.0  # Graphiques
seaborn>=0.12.2  # Visualisation statistique
plotly>=5.18.0  # Visualisation interactive

# Tests et Qualité
pytest>=7.4.0  # Framework de tests
pytest-asyncio>=0.21.0  # Tests asynchrones
pytest-mock>=3.11.0  # Mocking pour les tests
pytest-cov>=4.1.0  # Couverture de code
coverage>=7.3.0  # Rapport de couverture

# Formatage et Linting
black>=23.10.0  # Formattage de code
isort>=5.12.0  # Organisation des imports
flake8>=6.1.0  # Linting de code
mypy>=1.6.1  # Vérification de types
pre-commit>=3.3.0  # Hooks Git

# Déploiement
docker>=6.1.0  # SDK Docker
boto3>=1.34.0  # AWS SDK
python-json-logger>=2.0.7  # Logs JSON pour conteneurs

# Sécurité
cryptography>=41.0.0  # Chiffrement
python-keycloak>=3.5.0  # Authentification Keycloak
passlib[bcrypt]>=1.7.4  # Hachage de mots de passe
